ms-controller/requirements.txt
 
Flask==1.1.2
flask-cors==3.0.8
pymongo==3.9.0
python-bsonjs==0.2.0
cryptography
requests